cleanup apple clang 13.0.0 warnings
Clean up all warnings generated with `-Wall -Wextra` set.

Note: large whitespace changes in the amalgamated header files are due to
the use of quom to generate them.

closes #83

+++ include/tiled/Layer.hpp
@@ -508,7 +508,7 @@ void tson::Layer::createTileData(const Vector2i &mapSize, bool isInfiniteMap)
     {
         std::for_each(m_data.begin(), m_data.end(), [&](uint32_t tileId)
         {
-            if (x == mapSize.x)
+            if (static_cast<int>(x) == mapSize.x)
             {
                 ++y;
                 x = 0;
